Rename rpc_test folder to rpctest
