+++ api/rpctest/trace_test.go
